refactor(eval-workflows): 修正样本生成的领域归属
