+++ src/eval-workflows/sample-generation/generator.ts
@@ -1,13 +1,13 @@
 import { createExecutor } from '../../executors/index.js';
 import { executorSupportsSampleMocks } from '../../executors/core/capabilities.js';
-import { DEFAULT_EVALUATION_GATE_THRESHOLD as DEFAULT_GATE_THRESHOLD } from '../../eval-workflows/evaluation-defaults.js';
-import { sampleMockReferenceKeys } from '../../eval-workflows/inputs/sample-contract.js';
-import type { Sample, SampleProvenance } from '../../eval-workflows/inputs/contracts/sample.js';
-import { SampleSchema } from '../../eval-workflows/inputs/schemas/sample-set.js';
-import { detailedSchemaIssue } from '../../eval-workflows/inputs/schemas/error.js';
-import { MockSchema } from '../../eval-workflows/inputs/schemas/mock.js';
+import { DEFAULT_EVALUATION_GATE_THRESHOLD as DEFAULT_GATE_THRESHOLD } from '../evaluation-defaults.js';
+import { sampleMockReferenceKeys } from '../inputs/sample-contract.js';
+import type { Sample, SampleProvenance } from '../inputs/contracts/sample.js';
+import { SampleSchema } from '../inputs/schemas/sample-set.js';
+import { detailedSchemaIssue } from '../inputs/schemas/error.js';
+import { MockSchema } from '../inputs/schemas/mock.js';
 import type { ExecutorFn } from '../../executors/contracts/ports.js';
-import type { Assertion } from '../../eval-workflows/inputs/contracts/assertion.js';
+import type { Assertion } from '../inputs/contracts/assertion.js';
 import type { ObservationInboxItem } from '../../observability/contracts/inbox.js';
 
 const SYSTEM_PROMPT = `你是一个评测用例生成器。你的任务是根据用户提供的 skill（系统提示词）内容，生成高质量的评测用例。
